Make the serialized dictionary null-terminate its last element.
This behavior isn't required according to RFC5050, because the dictionary
is preceded by a dictionary length field so that a dictionary parser can
determine the end of the dictionary without a null terminator.

However, DTN2 does null terminate things so this code will do the same, at
least until it can be confirmed that this doesn't affect interoperability.