Add ComparisonStream, a test class that throws an exception if streams don't match.
ComparisonStream is an OutputStream that takes an InputStream as a constructor
argument.  Every time you write a byte to ComparisonStream, it reads a byte
from its InputStream, and compares the two.  If they match, nothing happens
(the written byte is thrown away).  If they don't match, an exception is
thrown.

This is a helpful class for testing the equality of two streams, especially
a bundle that is being serialized versus a bundle written to a file on disk.