Make comparisonStream better by adding finish() method.
Before this commit, there was no finish() method for the comparisonStream,
so it would erroneously not throw an IOException in the case that the InputStream
and OutputStreams matched for all the bytes up to len(OutputStream), but InputStream
had extra bytes after, i.e. it would say the following streams matched:

in:   "HEREISASTREAMWITHSOMEBYTESandsomebytesafterthat"
out:  "HEREISASTREAMWITHSOMEBYTES"

But now, if you call cmpstream.finish() it will verify that there are no more
bytes in the input stream.  It will throw a helpful IOException if there are.