Fix bug in bytesIntToBytes(): Handle zero properly.
When positive BigIntegers are converted to byte arrays, they will be
preceded with a zero byte if the MSB in the most significant byte is set,
to make it clear that they shouldn't be treated as negative twos
complement numbers.  But SDNVs are always nonnegative and prepending this
leading 0 byte is invalid.  So, bytesIntToBytes strips leading 0 bytes.

Before this commit, it would incorrectly handle the case of 0, where there
is exactly one byte and it is zeros.  It would consider this a "leading"
0 byte and try to access the next element of the array (OutOfBounds
exception).  This commit fixes that, and adds a test case.