Create skeleton for UDP Convergence Layer.
The UdpConvergenceLayer class is a Service.  Right now it only receives
bundles in UDP datagrams on the default port of 4556.  It prints those
bundles to the debug log and discards them (see UdpConvergenceLayerTest).

Next is going to be learning the Android IPC mechanisms so that the
UdpConvergenceLayer can prompt the rest of roboDTN when it receives a
bundle, and so that roboDTN can prompt UdpConvergenceLayer to send bundles.