Create new dialogs on button press to avoid using cached ones.
Before, Android was providing cached Date/TimePickerDialogs when a button
was clicked.  This meant that the date the button was showing may not match
the date the dialog would show, because the button caption is updated by
changing the SDNV.

Instead, we create a new dialog each time a button is pressed.